Fix errors with 'sudo reptyr -T'
When we attach a process using `reptyr -T`, we create a tempdir and
create a unix socket therein, which the target needs to be able to
connect() to. If `reptyr -T` is run as root, the directory will only be
available to root by default. chown the directory to the target, which
should work for both common cases of attaching a target of the same uid,
and root attaching a lesser uid's process.