use the emulator pid when chown'ing the socket
It's entirely possible the terminal emulator is running under a
different uid than the target process (e.g. someone ran `sudo`); In this
situation, we need the terminal emulator to be able to write to our
socket, not the victim.

We also need to read /proc/$pid/status instead of just stat()ing
/proc/$pid/status to deal with a terminal emulator that has dropped
privileges in the past (e.g. sshd), leaving its old credentials on the
inode.