FreeBSD: Fix tty stealing, enable at least that test (#114)
* freebsd: remove bogus double-open of procstat

Signed-off-by: Kyle Evans <kevans@FreeBSD.org>

* freebsd: record ctty of the victim

Signed-off-by: Kyle Evans <kevans@FreeBSD.org>

* freebsd: fix master-side discovery

Previously we were grabbing all master fds, rather than just those that
corresponded to the process we were targetting.  This blew up in the
tty-steal test as we dup'd right over our own ctty.

Signed-off-by: Kyle Evans <kevans@FreeBSD.org>

* Enable FreeBSD tests on Cirrus, but skip the basic test for now

We currently fail the basic test because TIOCNOTTY isn't actually
implemented on FreeBSD (yet). We can disable that for now and at least
ensure that the tty-stealing functionality doesn't regress.

Signed-off-by: Kyle Evans <kevans@FreeBSD.org>

* cirrus: drop CIRRUS_CLONE_DEPTH for faster clones

Signed-off-by: Kyle Evans <kevans@FreeBSD.org>