Prefer direct socket syscall APIs over `socketcall`
x86-32 requires the use of the `socketcall` syscall to access the socket
syscall interface, and the logic for handling `socketcall` assumes a
32-bit architecture. However, powerpc provides both the `socket` and
`socketcall` syscalls. The previous logic caused us to prefer
`socketcall` over `socket`, which ended poorly since powerpc is not a
32-bit architecture.

This reverses the order of precedence - we'll always use `socket` if
it's available, and only fall back to `socketcall` if it's not.