freebsd: add support for spilling syscall args to the stack
armv7 and powerpc both stash some number of arguments on the stack.  The
next commit will add armv7 support, so let's get the syscall setup
hashed out.

Stack args in the personality are indicated by having an offset above
the size of `struct reg`.  By convention, these should just use ~0UL
rather than sizeof(struct reg) + X to avoid issues if the struct reg
ABI changes.

There's a bit of a kludge here described in detail in the comments, but
the short version is that mmap's final argument is a 64-bit argument
that should be passed in as a pair of 32-bit ints.  We know the upper 32
bits will always be 0 because we don't use the offset argument, so we
just pad out to four arguments on the stack, with the excess being '0'.

Signed-off-by: Kyle Evans <kevans@FreeBSD.org>