fix(Makefile): Make bash completion honor PREFIX
The current version of the build system uses `pkgconfig` to discover the
location of the bash completion scripts. Typically, this script outputs
a directory under the `/usr` prefix. Unfortunately, if the user chooses
to override the build prefix with PREFIX, the BASHCOMPLETION variable is
not also updated.

In this commit, I propose to correct this issue by simply looking for
the `/usr` prefix and substituting it with the contents of `PREFIX`.
I've tested this approach using my `stow`-based "package" collection,
where I build commands with custom prefixes, and do not grant root or
sudo privileges when doing so. It appears to function normally for me.